+++ demo/test/components/previews/primer/alpha/auto_complete_preview.rb
@@ -2,21 +2,19 @@ module Primer
   module Alpha
     class AutoCompletePreview < ViewComponent::Preview
       def default
-        render(Primer::Beta::AutoComplete.new(label_text: "Select a fruit", input_id: "input-id", list_id: "test-id", src: "/auto_complete"))
+        render(Primer::Alpha::AutoComplete.new(label_text: "Select a fruit", input_id: "input-id", list_id: "test-id", src: "/auto_complete?version=alpha"))
       end
 
       def with_non_visible_label
-        render(Primer::Beta::AutoComplete.new(label_text: "Select a fruit", input_id: "input-id", list_id: "test-id", src: "/auto_complete", visually_hide_label: true))
+        render(Primer::Alpha::AutoComplete.new(label_text: "Select a fruit", input_id: "input-id", list_id: "test-id", src: "/auto_complete?version=alpha", is_label_visible: false))
       end
 
       def with_icon
-        render(Primer::Beta::AutoComplete.new(label_text: "Select a fruit", input_id: "input-id", list_id: "test-id", src: "/auto_complete")) do |c|
-          c.leading_visual_icon(icon: :search)
-        end
+        render(Primer::Alpha::AutoComplete.new(label_text: "Select a fruit", input_id: "input-id", list_id: "test-id", src: "/auto_complete?version=alpha", with_icon: true))
       end
 
-      def show_clear_button
-        render(Primer::Beta::AutoComplete.new(label_text: "Select a fruit", input_id: "input-id", list_id: "test-id", src: "/auto_complete", show_clear_button: true))
+      def with_clear_button
+        render(Primer::Alpha::AutoComplete.new(label_text: "Select a fruit", input_id: "input-id", list_id: "test-id", src: "/auto_complete?version=alpha", is_clearable: true))
       end
     end
   end
